requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
langchain-groq
langchain-core
python-dotenv
tiktoken
transformers
google-generativeai
requests
langchain
langchain-community
geopy
googlemaps
torch
folium
streamlit-folium
scipy
numpy
grpcio
pandas
pymupdf
streamlit
python-dotenv
torch
torchvision
Pillow
folium
streamlit-folium
geopy
googlemaps
faiss-cpu
sentence-transformers
langchain-groq
google-generativeai
pymupdf
tiktoken
transformers
